feat(psdl_ai_gen): extract OpenAI + Ollama PSDL synthesis into package
Third module extraction of the three-tier refactor (after psdl_meds,
psdl_vocab). Moves openai_service + ollama_service into an installable
psdl_ai_gen package; app/services/ shims re-export so callers
(routers/generate.py, routers/export.py) keep their imports.

Uses the flat package layout (package-dir={'psdl_ai_gen': '.'}) to avoid
the namespace-shadowing trap that the nested layout caused for psdl_vocab.

Verified: app.main boots from backend cwd; uvicorn serves 200; full
suite green (no new failures vs 73-passed baseline).
